# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "M1P9 Android",
    "description": "",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
    "/api/user/login": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "login",
        "tags": ["User"],
        "description": "Login",
        "parameters": [
          {
            "name": "Login",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Login success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Authentication failed",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "/api/user/signup": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "signup",
        "tags": ["User"],
        "description": "Signup",
        "parameters": [
          {
            "name": "Signup",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Signup"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Signup success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Email already used",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "/api/user/logout": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "logout",
        "tags": ["User"],
        "description": "Logout",
        "parameters": [
          {
            "name": "Logout",
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["id"],
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Logout success",
            "schema": {
              "$ref": "#/definitions/SuccessModel"
            }
          },
          "500": {
            "description": "Logout failed",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Login": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Signup": {
      "type": "object",
      "required": ["email", "password", "firstname", "lastname"],
      "properties": {
        "email": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "firstname": {
              "type": "string"
            },
            "lastname": {
              "type": "string"
            }
          }
        }
      }
    },
    "SuccessModel": {
      "type": "object",
      "required": ["success", "message"],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "ErrorModel": {
      "type": "object",
      "required": ["success", "message"],
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}